requirements.txt
discord.py==1.3.2
tabulate==0.8.6
SQLAlchemy==1.3.13
